Recstucturized the UI into separate files
