+++ src/constants.py
@@ -1,0 +1,1 @@
+ANN_MARGIN = 18   # pixels reserved above each text block for the annotation row
